Moved a lot of slideshow logic into Slideshow.cs
